에러 핸들링과 리팩토링입니다.

+++ server/routes/Favorite.js
@@ -4,35 +4,39 @@ var favorites = require("../models").Favorites;
 var router = express.Router();
 
 router.post("/toggle", (req, res) => {
-  favorites
-    .findOne({
-      where: {
-        user_id: req.body.user_id,
-        category_id: req.body.category_id
-      }
-    })
-    .then(result => {
-      if (!result) {
-        favorites
-          .create({
-            check: true,
-            user_id: req.body.user_id,
-            category_id: req.body.category_id
-          })
-          .then(result => res.send(JSON.stringify(result)))
-          .catch(err => res.send(JSON.stringify(err)));
-      } else {
-        favorites
-          .destroy({
-            where: {
+  if (req.token && req.token.id) {
+    favorites
+      .findOne({
+        where: {
+          user_id: req.body.user_id,
+          category_id: req.body.category_id
+        }
+      })
+      .then(result => {
+        if (!result) {
+          favorites
+            .create({
+              check: true,
               user_id: req.body.user_id,
               category_id: req.body.category_id
-            }
-          })
-          .then(result => res.send(JSON.stringify(result)))
-          .catch(err => res.send(JSON.stringify(err)));
-      }
-    });
+            })
+            .then(result => res.json(result))
+            .catch(err => res.json(err));
+        } else {
+          favorites
+            .destroy({
+              where: {
+                user_id: req.body.user_id,
+                category_id: req.body.category_id
+              }
+            })
+            .then(result => res.json(result))
+            .catch(err => res.json(err));
+        }
+      });
+  } else {
+    res.sendStatus(401);
+  }
 });
 
 module.exports = router;
